New Development: No             JIRA Issue:         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: No              Module(s): mstransform            Description: remove -Wsign-compare warning.
git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk@39487 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ code/mstransform/TVI/UVContSubTVI.cc
@@ -140,7 +140,7 @@ Bool UVContSubTVI::parseConfiguration(const Record &configuration)
 		if (nThreads_p > 1)
 		{
 #ifdef _OPENMP
-			if (omp_get_max_threads() < nThreads_p)
+			if (omp_get_max_threads() < (int)nThreads_p)
 			{
 				logger_p << LogIO::WARN << LogOrigin("UVContSubTVI", __FUNCTION__)
 						<< "Requested " <<  nThreads_p << " OMP threads but maximum possible is " << omp_get_max_threads()<< endl
